Errors in test reporting should be non-fatal.
Minor. Additions to git ignore.

Style. A boat-load of extra analysis options.

Documentation. Update changelog with the horrors of breaking changes.

Feature. Convenience methods for easyy access to reply and response status.

Bugfixes. Fix logic error in Response and Reply class and make sure
that the last line of the raw response is not an empty line.

Enhancement. Elaborate reponse tests.